requirements.txt
aiohttp==3.9.5
interactions.py==5.12.1
mariadb==1.1.10